Creating Properties

When a dot operator call getproperty or setproperty! it represent the property as a Symbol. Sometimes it's useful to have a method that retrieves a specific property so that users have a common API that doesn't change despite changes in the internal structure of types. @defprop automatically creates a method that calls getproperty and a method that calls setproperty!. This permits a sort of duck typing that can also be optimized for compile time performance. This section describes how to create properties quickly and some ways of customizing the behavior.

Simple property construction

FieldProperties.@defprop — Macro

@defprop Property{name}::Type block

Convient way of creating properties that wrap getproperty and setproperty! methods.

Examples

The simplest form simply creates a getter and setter.

julia> using FieldProperties

julia> @defprop Property1{:prop1}

julia> name(prop1)
:prop1

julia> proptype(prop1)
Any

julia> prop1
prop1 (generic function with 1 method)

julia> prop1!
prop1! (generic function with 1 method)

Define the propertie's type

julia> @defprop Property2{:prop2}::Int

julia> name(prop2)
:prop2

julia> proptype(prop2)
Int64

Define type requirement and default value.

julia> @defprop Property3{:prop3}::Int begin
           @getproperty x -> 1
       end

julia> name(prop3)
:prop3

julia> proptype(prop3)
Int64

julia> prop3(3)
1

Define a default value but no enforced return type but multiple getproperty methods.

julia> @defprop Property4{:prop4} begin
           @getproperty x::Int -> 1
           @getproperty x::String -> "1"
       end

julia> name(prop4)
:prop4

julia> proptype(prop4)
Any

julia> prop4(1)
1

julia> prop4
prop4 (generic function with 3 methods)

julia> prop4!
prop4! (generic function with 1 method)

Property types and interface

FieldProperties.AbstractProperty — Type

AbstractProperty{name}

A AbstractProperty serves as a common reference point for accessing methods related to name, where name is the dot accessed field (as in x.<name>).

See @defprop, @properties

Missing docstring.

Missing docstring for propname. Check Documenter's build log for details.

FieldProperties.proptype — Function

proptype(p, context) -> Type

Return the appropriate type for property p given context. This method allows unique type restrictions given different types for context.

FieldProperties.propconvert — Function

propconvert(p, context, v)

Ensures the value v is the appropriate type for property p given context. If v isn't the appropriate type then propconvert attempts to convert to the "correct type". The "correct type" is determined by proptype(p, x).

This is functionlly equivalent to convert(proptype(p, x), v).