Property Documentation and Reflection

A number of functions are provided for the purpose of documenting properties and manipulating them programmatically. These methods are still under development.

FieldProperties.propdoc — Function

propdoc(x) -> String

Returns documentation for property x.

FieldProperties.propdoclist — Function

propdoclist(ps...) -> String

Returns a markdown list of properties given multiple properties ps.